+++ src/Input.js
@@ -2,6 +2,7 @@ import {EventEmitter} from "../node_modules/djipevents/dist/djipevents.esm.min.j
 import {Message, WebMidi} from "./WebMidi.js";
 import {InputChannel} from "./InputChannel.js";
 import {Utilities} from "./Utilities.js";
+import {Forwarder} from "./Forwarder.js";
 
 /**
  * The `Input` class represents a single MIDI input port. This object is automatically instantiated
@@ -60,6 +61,8 @@ export class Input extends EventEmitter {
     this.channels = [];
     for (let i = 1; i <= 16; i++) this.channels[i] = new InputChannel(this, i);
 
+    this._forwarders = [];
+
     // Setup listeners
     this._midiInput.onstatechange = this._onStateChange.bind(this);
     this._midiInput.onmidimessage = this._onMidiMessage.bind(this);
@@ -76,6 +79,7 @@ export class Input extends EventEmitter {
     this.removeListener();
     this.channels.forEach(ch => ch.destroy());
     this.channels = [];
+    this._forwarders = [];
     if (this._midiInput) {
       this._midiInput.onstatechange = null;
       this._midiInput.onmidimessage = null;
@@ -175,6 +179,7 @@ export class Input extends EventEmitter {
    */
   _onMidiMessage(e) {
 
+
     // Create Message object from MIDI data
     const message = new Message(e.data);
 
@@ -225,6 +230,9 @@ export class Input extends EventEmitter {
     } else if (message.isChannelMessage) {   // channel messages
       this.channels[message.channel]._processMidiMessageEvent(event);
     }
+
+    // Forward message if forwarders have been defined
+    this._forwarders.forEach(forwarder => forwarder.forward(message));
 
   }
 
@@ -702,6 +710,61 @@ export class Input extends EventEmitter {
 
     }
 
+  }
+
+  /**
+   * Adds a forwarder that will forward all incoming MIDI messages matching the criteria to the
+   * specified output destination(s). This is akin to the hardware MIDI THRU port with the added
+   * benefit of being able to filter which data is forwarded.
+   *
+   * @param {Output|Output[]} destinations An [`Output`](Output) object, or an array of such objects,
+   * to forward messages to.
+   * @param {object} [options={}]
+   * @param {string|string[]} [options.types] A message type (`"noteon"`, `"controlchange"`, etc.),
+   * or an array of such types, that the message type must match in order to be forwarded. If this
+   * option is not specified, all types of messages will be forwarded. Valid messages are the ones
+   * found in either [`MIDI_SYSTEM_MESSAGES`](Enumerations#MIDI_SYSTEM_MESSAGES) or
+   * [`MIDI_CHANNEL_MESSAGES`](Enumerations#MIDI_CHANNEL_MESSAGES).
+   * @param {number} [options.channels=[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16]] A
+   * MIDI channel number or an array of channel numbers that the message must match in order to be
+   * forwarded. By default all MIDI channels are included (`1` to `16`).
+   *
+   * @returns {Forwarder} The [`Forwarder`](Forwarder) object created to handle the forwarding. This
+   * is useful if you wish to manipulate or remove the [`Forwarder`](Forwarder) later on.
+   */
+  addForwarder(output, options = {}) {
+
+    let forwarder;
+
+    // Unless 'output' is a forwarder, create a new forwarder
+    if (output instanceof Forwarder) {
+      forwarder = output;
+    } else {
+      forwarder = new Forwarder(output, options);
+    }
+
+    this._forwarders.push(forwarder);
+    return forwarder;
+
+  }
+
+  /**
+   * Removes the specified forwarder.
+   * @param {Forwarder} forwarder The [`Forwarder`](Forwarder) to remove (the
+   * [`Forwarder`](Forwarder) object is returned when calling `addForwarder()`.
+   */
+  removeForwarder(forwarder) {
+    this._forwarders = this._forwarders.filter(item => item !== forwarder);
+  }
+
+  /**
+   * Checks whether the specified forwarded has already been attached to this input.
+   * @param {Forwarder} forwarder The [`Forwarder`](Forwarder) to check (the
+   * [`Forwarder`](Forwarder) object is returned when calling `addForwarder()`.
+   * @returns {boolean}
+   */
+  hasForwarder(forwarder) {
+    return this._forwarders.includes(forwarder);
   }
 
   /**