+++ CHANGELOG.md
@@ -58,6 +58,9 @@ options of `WebMidi.enable()`.
 playing or not. It also has a new `notesState` property which is an array holding the playing status
 of all notes (0-127).
 
+- It is now possible to add a forwarder to an `Input` that will forward MIDI messages to a specified
+output. Also, the inbound messages can be filtered for forwarding by message type and channel.
+
 ### Changed
 
 - [BREAKING CHANGE] Passing `undefined` as the `channel` value to `addListener()` no longer means